+++ docs/releasenotes.rst
@@ -3,8 +3,8 @@ Release notes
 
 Develop
 -------
-- deprecate `open`, use `link` instead to instantiate Darr objects from disk
-  arra data
+- `open` is the canonical function to instantiate Darr objects from disk-based
+  array data; `link` is kept as a deprecated alias
 
 
 Version 0.6.3